requirements.txt
requests~=2.32.3
python-dotenv>=1.0.1
matplotlib~=3.9.2
fitbit~=0.3.1
pandas==2.2.3
